transalations added and iife created for cookie-banner to pick up cookie name from hidden field
