+++ app/steps/ui/static/cookies/template.html
@@ -199,7 +199,7 @@
                     ],
                     rows: [
                         [
-                            { text: content.identityCookieName1 | safe },
+                            { text: SECURITY_COOKIE | safe },
                             { text: content.identityCookiePurpose1 | safe },
                             { text: content.identityCookieExpiry1 | safe }
                         ]
