+++ app/views/includes/scripts_ga_1.html
@@ -1,7 +1,14 @@
 {% if globals.enableTracking == "true" %}
     <!-- Data Layer -->
     <script nonce="{{ globals.nonce }}">
-      dataLayer = [];
+      if (!dataLayer) {
+        dataLayer = [];
+      }
+      const userPreferenceCookie = document.cookie.match(new RegExp('(^| )' + 'cm-user-preferences' + '=([^;]+)'));
+      if(userPreferenceCookie) {
+        const cookiePreferences = JSON.parse(decodeURIComponent(factCookie[2]));
+        dataLayer.push({'event': 'Cookie Preferences', 'cookiePreferences': cookiePreferences});
+      }
     </script>
    <!-- Google Tag Manager -->
 <script nonce='{{ globals.nonce }}'>(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
